Add files via upload
This Dashboard is created as part of my learning experience while learning SAP Analytics Cloud. The dataset is related to Sales data which is then processed, Geo-Enriched, Transformed and created a story. Later a Dynamic Dashboard is created using the story which is interactive. Thus it is insightful for the business user to drive meaningful insights and Decision Making.